feat: use new contentful endpoint
