+++ package.json
@@ -12,7 +12,6 @@
     "@remix-run/react": "^1.7.4",
     "@remix-run/serve": "^1.7.4",
     "@urql/core": "^3.0.5",
-    "gatsby-plugin-image": "^2.24.0",
     "graphql": "^16.6.0",
     "isbot": "^3.5.4",
     "react": "^18.2.0",
