+++ utils/valhalla-client.js
@@ -1,6 +1,6 @@
 import { createClient } from '@urql/core';
 
 export const client = createClient({
-  url: 'https://valhallaexamples-prod.staging-valhalla-api.io/',
+  url: 'https://valhallasourceconfig123-prod.valhalla-api.io/graphql',
   requestPolicy: 'network-only'
 });
